# GameDebugClient/src/api/game.swagger.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "Rest Adventure - Game API",
    "version": "0.0.0-alpha.0.233+2a1139010fada7f6864f85e7bd96490563a587b1"
  },
  "servers": [
    {
      "url": "https://localhost:7056"
    }
  ],
  "paths": {
...
    "/game/team/characters/{characterGuid}/combats": {
      "get": {
        "tags": [
          "Combats"
        ],
        "summary": "Get combats",
        "operationId": "Combats_GetCombats",
        "parameters": [
          {
            "name": "characterGuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Combat"
                  }
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "api-key": []
          }
        ]
      }
    },
    "/game/team/characters/{characterGuid}/combats/{combatGuid}/history": {
      "get": {
        "tags": [
          "Combats"
        ],
        "summary": "Search combat history",
        "operationId": "CombatsHistory_SearchCombatHistory",
        "parameters": [
          {
            "name": "characterGuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          },
          {
            "name": "combatGuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 2
          },
          {
            "name": "PageNumber",
            "in": "query",
            "description": "The page number\n            ",
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 3
          },
          {
            "name": "PageSize",
            "in": "query",
            "description": "The page size\n            ",
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 4
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResultOfCombatHistoryEntry"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "api-key": []
          }
        ]
      }
    },
    "/game/team/characters/{characterGuid}/combats/archived": {
      "get": {
        "tags": [
          "Combats"
        ],
        "summary": "Search archived combats",
        "operationId": "CombatsHistory_SearchArchivedCombats",
        "parameters": [
          {
            "name": "characterGuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          },
          {
            "name": "PageNumber",
            "in": "query",
            "description": "The page number\n            ",
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 2
          },
          {
            "name": "PageSize",
            "in": "query",
            "description": "The page size\n            ",
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 3
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResultOfArchivedCombat"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "api-key": []
          }
        ]
      }
    },
    "/game/team/characters/{characterGuid}/combats/archived/{combatGuid}/history": {
      "get": {
        "tags": [
          "Combats"
        ],
        "summary": "Search archived combats history",
        "operationId": "CombatsHistory_SearchArchivedCombatHistory",
        "parameters": [
          {
            "name": "characterGuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          },
          {
            "name": "combatGuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 2
          },
          {
            "name": "PageNumber",
            "in": "query",
            "description": "The page number\n            ",
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 3
          },
          {
            "name": "PageSize",
            "in": "query",
            "description": "The page size\n            ",
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 4
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResultOfCombatHistoryEntry"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "api-key": []
          }
        ]
      }
    },
...
    "/game/content/characters/classes/{characterClassId}": {
      "get": {
        "tags": [
          "Game"
        ],
        "summary": "Get character class",
        "operationId": "GameContent_GetCharacterClass",
        "parameters": [
          {
            "name": "characterClassId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CharacterClass"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "api-key": []
          }
        ]
      }
    },
    "/game/content/items/{itemId}": {
      "get": {
        "tags": [
          "Game"
        ],
        "summary": "Get item",
        "operationId": "GameContent_GetItem",
        "parameters": [
          {
            "name": "itemId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Item"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "api-key": []
          }
        ]
      }
    },
    "/game/content/locations/{locationId}": {
      "get": {
        "tags": [
          "Game"
        ],
        "summary": "Get location",
        "operationId": "GameContent_GetLocation",
        "parameters": [
          {
            "name": "locationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Location"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "api-key": []
          }
        ]
      }
    },
    "/game/content/jobs/{jobId}": {
      "get": {
        "tags": [
          "Game"
        ],
        "summary": "Get job",
        "operationId": "GameContent_GetJob",
        "parameters": [
          {
            "name": "jobId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "api-key": []
          }
        ]
      }
    },
    "/game/content/static-objects/{staticObjectId}": {
      "get": {
        "tags": [
          "Game"
        ],
        "summary": "Get static object",
        "operationId": "GameContent_GetHarvestable",
        "parameters": [
          {
            "name": "staticObjectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StaticObject"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "api-key": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "StaticObject": {
        "type": "object",
        "description": "Static object\n            ",
        "additionalProperties": false,
        "required": [
          "id",
          "name"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "The unique ID of the static object\n            ",
            "format": "guid",
            "minLength": 1
          },
          "name": {
            "type": "string",
            "description": "The name of the static object\n            ",
            "minLength": 1
          },
          "description": {
            "type": "string",
            "description": "The description of the static object\n            ",
            "nullable": true
          }
        }
      },
...
      "ProblemDetails": {
        "type": "object",
        "additionalProperties": {
          "nullable": true
        },
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "Combat": {
        "type": "object",
        "discriminator": {
          "propertyName": "$phase",
          "mapping": {
            "preparation": "#/components/schemas/CombatInPreparation",
            "ongoing": "#/components/schemas/OngoingCombat"
          }
        },
        "description": "Combat base\n            ",
        "x-abstract": true,
        "additionalProperties": false,
        "required": [
          "id",
          "$phase"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "The unique ID of the combat instance\n            ",
            "format": "guid",
            "minLength": 1
          },
          "$phase": {
            "type": "string"
          }
        }
      },
      "CombatInPreparation": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Combat"
          },
          {
            "type": "object",
            "description": "Combat in preparation\n            ",
            "additionalProperties": false,
            "required": [
              "attackers",
              "attackersOptions",
              "defenders",
              "defendersOptions"
            ],
            "properties": {
              "attackers": {
                "type": "array",
                "description": "The attackers in the combat instance\n            ",
                "items": {
                  "$ref": "#/components/schemas/EntityMinimal"
                }
              },
              "attackersOptions": {
                "description": "The options of the attacker side\n            ",
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/CombatFormationOptions"
                  }
                ]
              },
              "defenders": {
                "type": "array",
                "description": "The defenders in the combat instance\n            ",
                "items": {
                  "$ref": "#/components/schemas/EntityMinimal"
                }
              },
              "defendersOptions": {
                "description": "The options of the defender side\n            ",
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/CombatFormationOptions"
                  }
                ]
              }
            }
          }
        ]
      },
...
      "OngoingCombat": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Combat"
          },
          {
            "type": "object",
            "description": "Combat instance\n            ",
            "additionalProperties": false,
            "required": [
              "turn",
              "attackers",
              "defenders"
            ],
            "properties": {
              "turn": {
                "type": "integer",
                "description": "The current turn of the combat instance\n            ",
                "format": "int32"
              },
              "attackers": {
                "type": "array",
                "description": "The attackers in the combat instance\n            ",
                "items": {
                  "$ref": "#/components/schemas/CombatEntity"
                }
              },
              "defenders": {
                "type": "array",
                "description": "The defenders in the combat instance\n            ",
                "items": {
                  "$ref": "#/components/schemas/CombatEntity"
                }
              }
            }
          }
        ]
      },
...
      "SearchResultOfCombatHistoryEntry": {
        "type": "object",
        "description": "Search result\n            ",
        "additionalProperties": false,
        "required": [
          "items",
          "pageNumber",
          "pageSize",
          "totalItemsCount",
          "totalPagesCount"
        ],
        "properties": {
          "items": {
            "type": "array",
            "description": "The items found by the query\n            ",
            "items": {
              "$ref": "#/components/schemas/CombatHistoryEntry"
            }
          },
          "pageNumber": {
            "type": "integer",
            "description": "The page number corresponding to the results that have been selected\n            ",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "description": "The page size used by the search\n            ",
            "format": "int32"
          },
          "totalItemsCount": {
            "type": "integer",
            "description": "The total number of items matching the query\n            ",
            "format": "int32"
          },
          "totalPagesCount": {
            "type": "integer",
            "description": "The total number of pages\n            ",
            "format": "int32"
          }
        }
      },
      "CombatHistoryEntry": {
        "type": "object",
        "discriminator": {
          "propertyName": "$type",
          "mapping": {
            "preparation-started": "#/components/schemas/CombatPreparationStartedHistoryEntry",
            "started": "#/components/schemas/CombatStartedHistoryEntry",
            "entity-joined": "#/components/schemas/CombatEntityJoinedHistoryEntry",
            "entity-left": "#/components/schemas/CombatEntityLeftHistoryEntry",
            "entity-attacked": "#/components/schemas/CombatEntityAttackedHistoryEntry",
            "entity-died": "#/components/schemas/CombatEntityDiedHistoryEntry",
            "ended": "#/components/schemas/CombatEndedHistoryEntry"
          }
        },
        "description": "Combat history entry\n            ",
        "additionalProperties": false,
        "required": [
          "tick",
          "turn",
          "$type"
        ],
        "properties": {
          "tick": {
            "type": "integer",
            "description": "The tick at which the event happened\n            ",
            "format": "int64"
          },
          "turn": {
            "type": "integer",
            "description": "The turn at which the event happened\n            ",
            "format": "int64"
          },
          "$type": {
            "type": "string"
          }
        }
      },
...
      "SearchResultOfArchivedCombat": {
        "type": "object",
        "description": "Search result\n            ",
        "additionalProperties": false,
        "required": [
          "items",
          "pageNumber",
          "pageSize",
          "totalItemsCount",
          "totalPagesCount"
        ],
        "properties": {
          "items": {
            "type": "array",
            "description": "The items found by the query\n            ",
            "items": {
              "$ref": "#/components/schemas/ArchivedCombat"
            }
          },
          "pageNumber": {
            "type": "integer",
            "description": "The page number corresponding to the results that have been selected\n            ",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "description": "The page size used by the search\n            ",
            "format": "int32"
          },
          "totalItemsCount": {
            "type": "integer",
            "description": "The total number of items matching the query\n            ",
            "format": "int32"
          },
          "totalPagesCount": {
            "type": "integer",
            "description": "The total number of pages\n            ",
            "format": "int32"
          }
        }
      },
...
      "CharacterClass": {
        "allOf": [
          {
            "$ref": "#/components/schemas/CharacterClassMinimal"
          },
          {
            "type": "object",
            "description": "Character class\n            ",
            "additionalProperties": false,
            "required": [
              "description",
              "levelCaps"
            ],
            "properties": {
              "description": {
                "type": "string",
                "description": "The description of the character class\n            ",
                "minLength": 1
              },
              "levelCaps": {
                "type": "array",
                "description": "The level caps of the character class\n            ",
                "items": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        ]
      },
      "Item": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ItemMinimal"
          },
          {
            "type": "object",
            "description": "Item\n            ",
            "additionalProperties": false,
            "properties": {
              "description": {
                "type": "string",
                "description": "The description of the item\n            ",
                "nullable": true
              }
            }
          }
        ]
      },
      "Location": {
        "allOf": [
          {
            "$ref": "#/components/schemas/LocationMinimal"
          },
          {
            "type": "object",
            "description": "Map location\n            ",
            "additionalProperties": false,
            "required": [
              "discovered",
              "connectedLocations"
            ],
            "properties": {
              "discovered": {
                "type": "boolean",
                "description": "Has this location been discovered by the player.\nIf false, the connected locations will be hidden.\n            "
              },
              "connectedLocations": {
                "type": "array",
                "description": "The locations connected to this one\n            ",
                "items": {
                  "$ref": "#/components/schemas/LocationMinimal"
                }
              }
            }
          }
        ]
      },
      "Job": {
        "allOf": [
          {
            "$ref": "#/components/schemas/JobMinimal"
          },
          {
            "type": "object",
            "description": "Job\n            ",
            "additionalProperties": false,
            "required": [
              "innate",
              "levelCaps",
              "harvests"
            ],
            "properties": {
              "description": {
                "type": "string",
                "description": "The description of the job\n            ",
                "nullable": true
              },
              "innate": {
                "type": "boolean",
                "description": "Is the job innate?\n            "
              },
              "levelCaps": {
                "type": "array",
                "description": "The experience to reach each level of the job\n            ",
                "items": {
                  "type": "integer",
                  "format": "int32"
                }
              },
              "harvests": {
                "type": "array",
                "description": "The harvests provided by the job\n            ",
                "items": {
                  "$ref": "#/components/schemas/HarvestableEntityHarvestMinimal"
                }
              }
            }
          }
        ]
      },
...
    },
    "securitySchemes": {
      "api-key": {
        "type": "apiKey",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
...
}